implemented TotalUndisposedAllocationCount
