+++ tests/ImageSharp.Tests/Memory/Allocators/MemoryDiagnosticsTests.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0.
 
 using System;
-using System.Buffers;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using Microsoft.DotNet.RemoteExecutor;
